requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.22
PyMuPDF==1.23.8
python-magic==0.4.27
python-magic-bin==0.4.14; sys_platform == 'win32'
nltk==3.6.3
uuid==1.30
